Add FLOAT4E2M1 support to relevant operators (#6283)
### Description
- FLOAT4E2M1 has been added to proto in
https://github.com/onnx/onnx/pull/6318
- This PR adds FLOAT4E2M1 support for QuantizeLinear, DequantizeLinear,
Cast, CastLike (opset 23).
- Also add support to non-compute ops: Constant, ConstantOfShape,
Identity, Reshape, Shape, Size, If, Loop, Scan, Flatten, Pad, Squeeze,
Unsqueeze, Transpose (opset 23).

Similar to INT4/UNIT4, FP4 weights/inputs are expected to be packed.

---------

+++ onnx/numpy_helper.py
@@ -234,8 +234,8 @@ def unpacked_float4e2m1_to_float32(x: npt.NDArray[np.uint8]) -> npt.NDArray[np.f
     """
     # x is stored in 4 LSB of int
     sign = np.where(np.bitwise_and(x, 0x08), -1, 1)
-    mantissa = x & 0x01
-    exponent = (x & 0x06) >> 1
+    mantissa = (x & 0x01).astype(np.float32)
+    exponent = ((x & 0x06) >> 1).astype(np.float32)
 
     val = np.where(
         exponent == 0,
